batch.py was updated to allow repeated data queries of the same name.
Reset.py was removed,
the reset definition was moved to new_catalogs.py.

Removed a large part of the star_names.py that is
no longer required.
